dependencies: Move dependency resolution into own package
- Update README format
- Move invocation under a single method instead of relying on caller to
  also handle progress event dispatch
