--- src/main/kotlin/com/open592/appletviewer/dependencies/DependencyType.kt
@@ -1,9 +1,0 @@
-package com.open592.appletviewer.dependencies
-
-/**
- * Represents the types of dependencies we will be resolving.
- */
-public enum class DependencyType {
-    BROWSERCONTROL,
-    LOADER,
-}
