+++ src/main/kotlin/com/open592/appletviewer/dependencies/resolver/DependencyType.kt
@@ -1,0 +1,6 @@
+package com.open592.appletviewer.dependencies.resolver
+
+public enum class DependencyType {
+    BROWSERCONTROL,
+    LOADER,
+}
